Use DN instead of counts everywhere (#338)
In line with sunpy/sunpy#7585, this replaces the usage of `ct/count`
with `DN` everywhere. This may cause some friction in terms of
compatibility with sunpy as prior to v6, the units of any AIA map are
converted to `ct` rather than `DN`. I'm not sure the best way to resolve
that. We could of course just pin to >= v6, but that seems aggressive.

---------

Co-authored-by: Nabil Freij <nabil.freij@gmail.com>