Add function for interpreting quality flag (#350)
The `QUALITY` keyword in the AIA headers is a 32-bit integer that
denotes the quality of an observation. Flags associated with the
observation quality are encoded bit-wise, with a different message
associated with each bit flag. See [section 7.7.6 of the SDO user
guide](https://www.lmsal.com/sdodocs/doc/dcur/SDOD0060.zip/zip/entry/)
for more details.

This PR adds a function that takes in the value of the `QUALITY` key and
returns a list of messages associated with any quality flags that are
set.